requirements.txt
streamlit
langchain
boto3
python-dotenv
pytest
black
flake8